+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Use Shift instead of Alt for Videos/Channels page shortcuts
 - Remove old YouTube Email Notifier migration
 
 ## 1.7.1 - 2024 Jan 28