+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next
+## 1.8.0 - 2025 Sep 9
 - Use Shift instead of Alt for Videos/Channels page shortcuts
 - Add option to disable window decorations
 - Fix error when a channel has scheduled livestreams